Extract btsieve runner into own file
